+++ lib/pages/chat/chat_input_row.dart
@@ -342,6 +342,10 @@ class ChatInputRow extends StatelessWidget {
                             ),
                             onLongPress: () => recordingViewModel
                                 .startRecording(controller.room),
+                            style: IconButton.styleFrom(
+                              backgroundColor: theme.bubbleColor,
+                              foregroundColor: theme.onBubbleColor,
+                            ),
                             icon: const Icon(Icons.mic_none_outlined),
                           )
                         : IconButton(