+++ lib/widgets/theme_builder.dart
@@ -13,17 +13,20 @@ class ThemeBuilder extends StatefulWidget {
     ThemeMode themeMode,
     Color? primaryColor,
     bool pureBlack,
+    bool twemoji,
   ) builder;
 
   final String themeModeSettingsKey;
   final String primaryColorSettingsKey;
   final String pureBlackSettingsKey;
+  final String twemojiSettingsKey;
 
   const ThemeBuilder({
     required this.builder,
     this.themeModeSettingsKey = 'theme_mode',
     this.primaryColorSettingsKey = 'primary_color',
     this.pureBlackSettingsKey = 'pure_black',
+    this.twemojiSettingsKey = 'xyz.extera.next.twemojiFont',
     super.key,
   });
 
@@ -36,12 +39,15 @@ class ThemeController extends State<ThemeBuilder> {
   ThemeMode? _themeMode;
   Color? _primaryColor;
   bool? _pureBlack;
+  bool? _twemoji;
 
   ThemeMode get themeMode => _themeMode ?? ThemeMode.system;
 
   Color? get primaryColor => _primaryColor;
 
   bool get pureBlack => _pureBlack ?? false;
+
+  bool get twemoji => _twemoji ?? false;
 
   static ThemeController of(BuildContext context) =>
       Provider.of<ThemeController>(
@@ -56,12 +62,14 @@ class ThemeController extends State<ThemeBuilder> {
     final rawThemeMode = preferences.getString(widget.themeModeSettingsKey);
     final rawColor = preferences.getInt(widget.primaryColorSettingsKey);
     final rawPureBlack = preferences.getBool(widget.pureBlackSettingsKey);
+    final rawTwemoji = preferences.getBool(widget.twemojiSettingsKey);
 
     setState(() {
       _themeMode = ThemeMode.values
           .singleWhereOrNull((value) => value.name == rawThemeMode);
       _primaryColor = rawColor == null ? null : Color(rawColor);
       _pureBlack = rawPureBlack;
+      _twemoji = rawTwemoji;
     });
   }
 
@@ -99,6 +107,15 @@ class ThemeController extends State<ThemeBuilder> {
     });
   }
 
+  Future<void> setTwemoji(bool newTwemoji) async {
+    final preferences =
+        _sharedPreferences ??= await SharedPreferences.getInstance();
+    await preferences.setBool(widget.twemojiSettingsKey, newTwemoji);
+    setState(() {
+      _twemoji = newTwemoji;
+    });
+  }
+
   @override
   void initState() {
     WidgetsBinding.instance.addPostFrameCallback(_loadData);
@@ -115,6 +132,7 @@ class ThemeController extends State<ThemeBuilder> {
           themeMode,
           primaryColor ?? light?.primary,
           pureBlack,
+          twemoji,
         ),
       ),
     );