+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - Fixed ringtone on Android.
 - Brought back image previews before sending.
 - Added a list of privacy settings set for different chats.
+- Added update checking.
+- Added call button in profile view.
 
 ## Extera 25.1.1
 - Added context menu for messages. Now, when selecting a single message, a context menu will appear. Multi-selection is still available.